move functionality to utils.conversion
