+++ Wrappers/Python/test/test_hdf5.py
@@ -4,11 +4,8 @@ import unittest
 import h5py
 import numpy as np
 import vtk
-from ccpi.viewer.utils.conversion import Converter, cilHDF5ResampleReader, cilHDF5CroppedReader
+from ccpi.viewer.utils.conversion import *
 from ccpi.viewer.utils.hdf5_io import (HDF5Reader, HDF5SubsetReader, write_image_data_to_hdf5)
-
-
-from .utils import calculate_target_downsample_shape
 
 
 class TestHDF5IO(unittest.TestCase):
