+++ Wrappers/Python/test/test_io.py
@@ -4,12 +4,8 @@ import unittest
 import h5py
 import numpy as np
 import vtk
-from ccpi.viewer.utils.conversion import Converter
+from ccpi.viewer.utils.conversion import *
 from ccpi.viewer.utils.io import ImageReader, cilviewerHDF5Writer, cilviewerHDF5Reader
-
-
-from .utils import calculate_target_downsample_shape
-
 
 class TestImageReaderAndWriter(unittest.TestCase):
